- improved UI - fix: m_ThreadExceptionHandlers should be synchronized
git-svn-id: http://crashrpt.googlecode.com/svn@523 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/1.2/src/crashrptprobe/src/MinidumpReader.cpp
@@ -602,7 +602,7 @@ BOOL CALLBACK ReadProcessMemoryProc64(
         return FALSE;
 
       *lpNumberOfBytesRead = (DWORD)lBytesRead;
-      memcpy(lpBuffer, (LPBYTE)mr.m_pStartPtr+dwOffs, lBytesRead);
+      memcpy(lpBuffer, (LPBYTE)mr.m_pStartPtr+dwOffs, (size_t)lBytesRead);
      
       return TRUE;
     }
